Refactor UserDetailsService and remove debug properties
The main action in this commit is the refactoring of the UserDetailsServiceImpl by relocating it to greencity.security.service package from the controller package. Additionally, properties related to debugging and circular-references have been removed in all the application environment properties files.
