+++ core/src/main/resources/application-dev.properties
@@ -63,5 +63,3 @@ greencity.server.address = http://localhost:8080
 
 #Swagger
 springdoc.swagger-ui.doc-expansion=none
-logging.level.org.springframework.web=DEBUG
-spring.main.allow-circular-references=true
