+++ core/src/main/resources/application-docker.properties
@@ -59,5 +59,3 @@ spring.thymeleaf.cache=false
 
 #RestTemplate
 greencity.server.address = http://core:8080
-logging.level.org.springframework.web=DEBUG
-#spring.main.allow-circular-references=true
