+++ core/src/main/resources/application-prod.properties
@@ -68,6 +68,3 @@ greencity.server.address = ${GREENCITY_SERVER_ADDRESS}
 
 #Thymeleaf
 spring.resources.static-locations=classpath:/static/
-logging.level.org.springframework.web=DEBUG
-#spring.main.allow-circular-references=true
-
