+++ core/src/main/resources/application.properties
@@ -1,4 +1,2 @@
 spring.profiles.active=${PROFILE:dev}
-logging.level.org.springframework.web=DEBUG
 springdoc.swagger-ui.doc-expansion=none
-#spring.main.allow-circular-references=true
